@@ -1,81 +1,59 @@
-import React, { useEffect } from "react";
-import "./App.css";
 import Header from "./Header";
+import './App.css';
+// import Orders from "./Orders";
 import Home from "./Home";
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import Checkout from "./Checkout";
 import Login from "./Login";
-import Payment from "./Payment";
-import Orders from "./Orders";
-import { auth } from "./firebase";
-import { useStateValue } from "./StateProvider";
-import { loadStripe } from "@stripe/stripe-js";
-import { Elements } from "@stripe/react-stripe-js";
+import Checkout from "./Checkout";
 import Buy from "./Buy";
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  
+} from "react-router-dom";
 
-const promise = loadStripe(
-  "pk_test_51HPvU9DFg5koCdLGJJbNo60QAU99BejacsvnKvT8xnCu1wFLCuQP3WBArscK3RvSQmSIB3N0Pbsc7TtbQiJ1vaOi00X9sIbazL"
-);
+
 
 function App() {
-  const [{}, dispatch] = useStateValue();
-
-  useEffect(() => {
-    // will only run once when the app component loads...
-
-    auth.onAuthStateChanged((authUser) => {
-      console.log("THE USER IS >>> ", authUser);
-
-      if (authUser) {
-        // the user just logged in / the user was logged in
-
-        dispatch({
-          type: "SET_USER",
-          user: authUser,
-        });
-      } else {
-        // the user is logged out
-        dispatch({
-          type: "SET_USER",
-          user: null,
-        });
-      }
-    });
-  }, []);
 
   return (
-    <Router>
-      <div className="app">
-        <Switch>
-          <Route path="/orders">
+
+    <>
+      <Router>
+      <div className ="app">
+      <Switch>
+          {/* <Route path="/orders">
             <Header />
             <Orders />
           </Route>
-          <Route path="/buy">
-            <Header />
-            <Buy/>
-          </Route>
-          <Route path="/login">
+      
+       
+          */}
+              <Route path="/login">
             <Login />
           </Route>
-          <Route path="/checkout">
-            <Header />
-            <Checkout />
+          <Route path="/buy">
+          <Header />
+          
+              <Buy/>
+           
           </Route>
-          <Route path="/payment">
+             <Route path="/checkout">
             <Header />
-            <Elements stripe={promise}>
-              <Payment />
-            </Elements>
-          </Route>
+           <Checkout />
+           </Route>
           <Route path="/">
             <Header />
             <Home />
           </Route>
+          
         </Switch>
       </div>
-    </Router>
-  );
-}
 
+      </Router>
+    </>
+
+  );
+
+}
 export default App;
