@@ -6,13 +6,14 @@ const Buy = () => {
         <div>
            
         
-            {/* <script>
-            function myFunction() {
+            {/* <Buy
+            function myFunctBuy
                 alert("you will 10 some coins after the product delivery")
             }
             </script> */}
             <h1>your order has been Successfully change,thanks for shoping with us..</h1>
-            <h2>you will 10 some coins after the product delivery</h2>
+            <h2>you will be created 10 coins in your account after the product delivery...</h2>
+            <script src="https://code.iconify.design/2/2.0.3/iconify.min.js"></script>
       </div>
     
         
