@@ -4,69 +4,70 @@ import Product from "./Product";
 import ModeStandbyRoundedIcon from '@mui/icons-material/ModeStandbyRounded';
 import { Stocks } from "./Stocks";
 
-
 function Home() {
   return (
     <div className="home">
-      <div className="home__container">
+    <div className="home__container">
+    
+
+      <div className="home__row">
       
+      <ModeStandbyRoundedIcon/>
+        <Product
+          id="12321341"
+          title="The Lean Startup: How Constant Innovation Creates Radically Successful Businesses Paperback"
+          price={11.96}
+          
+          image="https://images-na.ssl-images-amazon.com/images/I/51Zymoq7UnL._SX325_BO1,204,203,200_.jpg"
+        />
+       <Product
+          id="3254354345"
+          title="New Apple iPad Pro (12.9-inch, Wi-Fi, 128GB) - Silver (4th Generation)"
+          price={598.99}
+         
+          image="https://images-na.ssl-images-amazon.com/images/I/816ctt5WV5L._AC_SX385_.jpg"
+        />
+      </div>
 
-        <div className="home__row">
-        
-        <ModeStandbyRoundedIcon/>
-          <Product
-            id="12321341"
-            title="The Lean Startup: How Constant Innovation Creates Radically Successful Businesses Paperback"
-            price={11.96}
-            
-            image="https://images-na.ssl-images-amazon.com/images/I/51Zymoq7UnL._SX325_BO1,204,203,200_.jpg"
-          />
-         <Product
-            id="3254354345"
-            title="New Apple iPad Pro (12.9-inch, Wi-Fi, 128GB) - Silver (4th Generation)"
-            price={598.99}
-           
-            image="https://images-na.ssl-images-amazon.com/images/I/816ctt5WV5L._AC_SX385_.jpg"
-          />
-        </div>
+      <div className="home__row">
+        <Product
+          id="4903850"
+          title="Samsung "
+          price={199.99}
+          
+          image="https://images-na.ssl-images-amazon.com/images/I/71Swqqe7XAL._AC_SX466_.jpg"
+        />
+        <Product
+          id="23445930"
+          title="Amazon Echo (3rd generation) | Smart speaker with Alexa, Charcoal Fabric"
+          price={98.99}
+          
+          image="https://media.very.co.uk/i/very/P6LTG_SQ1_0000000071_CHARCOAL_SLf?$300x400_retinamobilex2$"
+        />
+        <Product
+          id="3254354345"
+          title="New Apple iPad Pro (12.9-inch, Wi-Fi, 128GB) - Silver (4th Generation)"
+          price={598.99}
+         
+          image="https://images-na.ssl-images-amazon.com/images/I/816ctt5WV5L._AC_SX385_.jpg"
+        />
+      </div>
 
-        <div className="home__row">
-          <Product
-            id="4903850"
-            title="Samsung "
-            price={199.99}
-            
-            image="https://images-na.ssl-images-amazon.com/images/I/71Swqqe7XAL._AC_SX466_.jpg"
-          />
-          <Product
-            id="23445930"
-            title="Amazon Echo (3rd generation) | Smart speaker with Alexa, Charcoal Fabric"
-            price={98.99}
-            
-            image="https://media.very.co.uk/i/very/P6LTG_SQ1_0000000071_CHARCOAL_SLf?$300x400_retinamobilex2$"
-          />
-          <Product
-            id="3254354345"
-            title="New Apple iPad Pro (12.9-inch, Wi-Fi, 128GB) - Silver (4th Generation)"
-            price={598.99}
-           
-            image="https://images-na.ssl-images-amazon.com/images/I/816ctt5WV5L._AC_SX385_.jpg"
-          />
-        </div>
-
-        <div className="home__row">
-          <Product
-            id="90829332"
-            title="Samsung - Super Ultra Wide Dual WQHD 5120 x 1440"
-            price={1094.98}
-            rating={4}
-            image="https://images-na.ssl-images-amazon.com/images/I/71Swqqe7XAL._AC_SX466_.jpg"
-          />
-          <Stocks />
-        </div>
+      <div className="home__row">
+        <Product
+          id="90829332"
+          title="Samsung - Super Ultra Wide Dual WQHD 5120 x 1440"
+          price={1094.98}
+          rating={4}
+          image="https://images-na.ssl-images-amazon.com/images/I/71Swqqe7XAL._AC_SX466_.jpg"
+        />
+        <Stocks />
       </div>
     </div>
+  </div>
   );
 }
 
 export default Home;
+
+
