@@ -1,38 +1,38 @@
 export const stockData = [
-  {
-    company: "The Lean Startup: How Constant Innovation Creates Radically Successful Businesses Paperback",
-    ticker: "TWTR",
-    stockPrice: "22.76 USD"
- 
-  },
-  {
-    company: "Kenwood kMix Stand Mixer for Baking, Stylish Kitchen Mixer with K-beater, Dough Hook and Whisk, 5 Litre Glass Bowl",
-    ticker: "SQ",
-    stockPrice: "22.76 USD"
+    {
+      company: "The Lean Startup: How Constant Innovation Creates Radically Successful Businesses Paperback",
+      ticker: "TWTR",
+      stockPrice: "22.76 USD"
    
-  },
-  {
-    company: "Samsung LC49RG90SSUXEN 49' Curved LED Gaming Monitor",
-    ticker: "SHOP",
-    stockPrice: "22.76 USD"
-   
-  },
-  {
-    company: "Amazon Echo (3rd generation) | Smart speaker with Alexa, Charcoal Fabric",
-    ticker: "RUN",
-    stockPrice: "9.87 USD"
-   
-  },
-  {
-    company: "New Apple iPad Pro (12.9-inch, Wi-Fi, 128GB) - Silver (4th Generation)",
-    ticker: "ADBE",
-    stockPrice: "300.99 USD"
-   
-  },
-  {
-    company: "Samsung LC49RG90SSUXEN 49' Curved LED Gaming Monitor - Super Ultra Wide Dual WQHD 5120 x 1440",
-    ticker: "HUBS",
-    stockPrice: "115.22 USD",
-   
-  }
-];
+    },
+    {
+      company: "Kenwood kMix Stand Mixer for Baking, Stylish Kitchen Mixer with K-beater, Dough Hook and Whisk, 5 Litre Glass Bowl",
+      ticker: "SQ",
+      stockPrice: "22.76 USD"
+     
+    },
+    {
+      company: "Samsung LC49RG90SSUXEN 49' Curved LED Gaming Monitor",
+      ticker: "SHOP",
+      stockPrice: "22.76 USD"
+     
+    },
+    {
+      company: "Amazon Echo (3rd generation) | Smart speaker with Alexa, Charcoal Fabric",
+      ticker: "RUN",
+      stockPrice: "9.87 USD"
+     
+    },
+    {
+      company: "New Apple iPad Pro (12.9-inch, Wi-Fi, 128GB) - Silver (4th Generation)",
+      ticker: "ADBE",
+      stockPrice: "300.99 USD"
+     
+    },
+    {
+      company: "Samsung LC49RG90SSUXEN 49' Curved LED Gaming Monitor - Super Ultra Wide Dual WQHD 5120 x 1440",
+      ticker: "HUBS",
+      stockPrice: "115.22 USD",
+     
+    }
+  ];
